Move contextFromString out of JaegerSpanContext into TextMapCodec
